feat:Added atribute in dml
Added extra atribute in dml in the "File" class.
The name of the atribute is "isDeleted" and its purpose is to verify if the file has be deleted since java doesn't let you delete objects but insteads garbage collects them, and we can't force garbage collector(we can but it isn't a good idea).

Closes #40